Refactor to improve testability of 'identify' subcommand
